Removal of no longer important usage of doubles.
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/branches/audiofloatbranch@399 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOSoundEngine.h
@@ -108,7 +108,7 @@ private:
 	std::vector<GOSamplerEntry>   m_Tremulants;
 
 	/* Per sampler decode buffers */
-	double                        m_FinalBuffer[GO_SOUND_BUFFER_SIZE];
+	float                         m_FinalBuffer[GO_SOUND_BUFFER_SIZE];
 	float                         m_VolumeBuffer[GO_SOUND_BUFFER_SIZE];
 
 	/* samplerGroupID:
